+++ src/events/BoundaryEvent.js
@@ -46,10 +46,15 @@ proto.execute = function execute(executeMessage) {
   const {isRootScope, executionId} = executeMessage.content;
 
   const eventDefinitionExecution = this[kExecution];
-  if (isRootScope) {
+  if (isRootScope && executeMessage.content.id === this.id) {
     this[kExecuteMessage] = executeMessage;
 
     const broker = this.broker;
+    if (executeMessage.fields.routingKey === 'execute.bound.completed') {
+      this._stop();
+      return broker.publish('execution', 'execute.completed', executeMessage.content, executeMessage.properties);
+    }
+
     const consumerTag = `_bound-listener-${executionId}`;
     this.attachedTo.broker.subscribeTmp('event', 'activity.leave', this._onAttachedLeave.bind(this), {
       noAck: true,
@@ -95,7 +100,7 @@ proto._onExecutionMessage = function onExecutionMessage(routingKey, message) {
 proto._onCompleted = function onCompleted(_, {content}) {
   if (!this.cancelActivity && !content.cancelActivity) {
     this._stop();
-    return this.broker.publish('execution', 'execute.completed', cloneContent(content, {cancelActivity: false}));
+    return this.broker.publish('execution', 'execute.completed', cloneContent(content, {isDefinitionScope: false, cancelActivity: false}));
   }
 
   this[kCompleteContent] = content;