+++ test/helpers/JavaScripts.js
@@ -59,7 +59,7 @@ function JavaScript(language, filename, scriptBody, environment) {
 
 JavaScript.prototype.execute = function execute(executionContext, callback) {
   const timers = this.environment.timers.register(executionContext);
-  return this.script.runInNewContext({...executionContext, ...timers, next: callback});
+  return this.script.runInNewContext({...executionContext, ...timers, next: callback, console});
 };
 
 function DummyScript(language, filename, logger) {